now data showing

+++ src/views/Dashboard/EmployeesView.vue
@@ -70,6 +70,16 @@
                                     <td class="px-4 py-2 font-medium text-gray-900 whitespace-nowrap dark:text-white"> {{
                                         employee.created_at }} </td>
                                     <td class="px-4 py-2 font-medium text-gray-900 whitespace-nowrap dark:text-white">
+                                        <button type="button" @click="showEmployee(employee.id)">
+                                            <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24"
+                                                stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
+                                                <path stroke-linecap="round" stroke-linejoin="round"
+                                                    d="M2.036 12.322a1.012 1.012 0 010-.639C3.423 7.51 7.36 4.5 12 4.5c4.638 0 8.573 3.007 9.963 7.178.07.207.07.431 0 .639C20.577 16.49 16.64 19.5 12 19.5c-4.638 0-8.573-3.007-9.963-7.178z" />
+                                                <path stroke-linecap="round" stroke-linejoin="round"
+                                                    d="M15 12a3 3 0 11-6 0 3 3 0 016 0z" />
+                                            </svg>
+                                        </button>
+
                                         <button>
                                             <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24"
                                                 stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
@@ -78,7 +88,7 @@
                                             </svg>
 
                                         </button>
-                                        <button>
+                                        <button @click="deleteEmployee(employee.id)">
                                             <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24"
                                                 stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
                                                 <path stroke-linecap="round" stroke-linejoin="round"
@@ -153,8 +163,9 @@
 
 
 
+        <h3 v-if="isShowModalOpen==true">This is modal </h3>
 
-        <!-- Main modal -->
+        <!-- employee create modal -->
         <div id="defaultModal" tabindex="-1" aria-hidden="true"
             class="hidden overflow-y-auto overflow-x-hidden fixed top-0 right-0 left-0 z-50 justify-center items-center w-full md:inset-0 h-modal md:h-full">
             <div class="relative p-4 w-full max-w-2xl h-full md:h-auto">
@@ -216,10 +227,8 @@
                                 <select v-model="newEmployee.department_id" id="category"
                                     class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500">
                                     <option selected="">Select category</option>
-                                    <option value="1">TV/Monitors</option>
-                                    <option value="2">PC</option>
-                                    <option value="3">Gaming/Console</option>
-                                    <option value="4">Phones</option>
+                                    <option v-for="department in departments" :key="department.id"
+                                        value="`{{ department.id }}}`">{{ department.name }}</option>
                                 </select>
                             </div>
                         </div>
@@ -244,24 +253,6 @@
                                 </li>
                             </div>
 
-
-                            <!-- <label for="Achievements"
-                                class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Achievements</label>
-                            <div class="grid grid-cols-4 gap-2">
-                                <li v-for="achievement in achievements" :key="achievement.id">
-                                    <div class="flex items-center ps-2 rounded hover:bg-gray-100 dark:hover:bg-gray-600">
-                                        <label class="w-full py-2 ms-2 text-sm font-medium text-gray-900 rounded dark:text-gray-300">
-                                            {{ achievement.name }}
-                                        </label>
-                                    </div>
-                                </li>
-                            </div> -->
-
-
-
-
-
-
                         </ul>
                         <button type="submit"
                             class="text-white inline-flex items-center bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">
@@ -272,6 +263,7 @@
                 </div>
             </div>
         </div>
+
 
 
     </AppLayout>
@@ -298,26 +290,28 @@ const newEmployee = ref(
 
 
 
-// Function to add an achievement to the newEmployee object
-// const addAchievement = (achievementName) => {
-//     newEmployee.value.achievements.push(achievementName);
-// };
-
-// // Function to remove an achievement from the newEmployee object
-// const removeAchievement = (achievementName) => {
-//     newEmployee.value.achievements = newEmployee.value.achievements.filter(
-//         (achievement) => achievement !== achievementName
-//     );
-// };
+// for show modal
+const isShowModalOpen = ref(false);
 
 
+// for store data 
 const employees = ref([]);
+const departments = ref([]);
 const achievements = ref([]);
 const token = localStorage.getItem('token');
 
 
 
 
+
+const fetchDepartments = async () => {
+    const response = await axios.get('http://127.0.0.1:8000/api/v1/backend/departments', {
+        headers: { Authorization: `Bearer ${token}` },
+    });
+    departments.value = response.data.departments;
+
+    // console.log(achievements);
+};
 
 const fetchAchievements = async () => {
     const response = await axios.get('http://127.0.0.1:8000/api/v1/backend/achievements', {
@@ -338,23 +332,13 @@ const fetchEmployees = async () => {
 };
 
 
-
-
-
-
 const createEmployee = async () => {
-
-
-    // console.log(newEmployee.value);
 
     const response = await axios.post('http://127.0.0.1:8000/api/v1/backend/employees', newEmployee.value, {
         headers: { Authorization: `Bearer ${token}` },
     });
 
-
-
-
-    console.log(response.data.employee);
+    // console.log(response.data.employee);
 
     newEmployee.value = {
         name: '',
@@ -368,16 +352,32 @@ const createEmployee = async () => {
 };
 
 const deleteEmployee = async (employeeId) => {
-    await axios.delete(`http://127.0.0.1:8000/api/v1/backend/employees/${employeeId}`, {
+
+    if (confirm('Are you sure?')) {
+        const response = await axios.delete(`http://127.0.0.1:8000/api/v1/backend/employees/${employeeId}`, {
+            headers: { Authorization: `Bearer ${token}` },
+        });
+        console.log(response.data.message);
+        fetchEmployees();
+    }
+
+};
+const showEmployee = async (employeeId) => {
+
+    const response = await axios.get(`http://127.0.0.1:8000/api/v1/backend/employees/${employeeId}`, {
         headers: { Authorization: `Bearer ${token}` },
     });
-    fetchEmployees(); // Refresh the employee list
+
+    isShowModalOpen.value = true;
+    console.log(isShowModalOpen);
+
 };
 
 // Fetch employees when the component is mounted
 onMounted(async () => {
     await fetchEmployees();
     await fetchAchievements();
+    await fetchDepartments();
 });
